refactoring TopicArticleVisualElement to more generic FormikVisualElement
