+++ src/containers/EditSubjectFrontpage/components/SubjectpageAbout.tsx
@@ -10,7 +10,7 @@ import { injectT } from '@ndla/i18n';
 import { FieldProps } from 'formik';
 import { SubjectpageEditType, TranslateType } from '../../../interfaces';
 import FormikField from '../../../components/FormikField';
-import TopicArticleVisualElement from '../../TopicArticlePage/components/TopicArticleVisualElement';
+import TopicArticleVisualElement from '../../FormikForm/components/FormikVisualElement';
 import PlainTextEditor from '../../../components/SlateEditor/PlainTextEditor';
 
 interface Props {
